Release 6 – add delact

+++ vmafia.py
@@ -245,8 +245,8 @@ def triggers(msg):
 
 @bot.message_handler(content_types=["new_chat_members"])
 def triggers(msg):
-    if msg.chat.id == GROUP_ID:
-        if not msg.new_chat_member.is_bot == True:
+    if not msg.new_chat_member.is_bot == True:
+        if msg.chat.id == GROUP_ID:
             uid = msg.new_chat_member.id
             keyboard = types.InlineKeyboardMarkup()
             url_button = types.InlineKeyboardButton(text="Читати правила", url="https://t.me/mafia_pravyla")
@@ -270,14 +270,17 @@ def triggers(msg):
             cur.execute("UPDATE messages SET id = %s WHERE type = %s", [new_bot_message_id, type_name])
             conn.commit()
             ##### Зберегти id поста
-    elif msg.chat.id == GROUP_ID_ACTIVE:
-        uid_admin = msg.new_chat_member.id
-        cur.execute("INSERT INTO active_admins (uids_admins) VALUES (%s)", [uid_admin])
-        conn.commit()
-        uids_admins.append(uid_admin)
+        elif msg.chat.id == GROUP_ID_ACTIVE:
+            if not uid_admins in uids_admins:
+                uid_admin = msg.new_chat_member.id
+                cur.execute("INSERT INTO active_admins (uids_admins) VALUES (%s)", [uid_admin])
+                conn.commit()
+                uids_admins.append(uid_admin)
+        else:
+            bot.send_message(msg.chat.id, text.notmafia.format(msg.from_user.id, msg.from_user.first_name),
+                         parse_mode="HTML")
     else:
-        bot.send_message(msg.chat.id, text.notmafia.format(msg.from_user.id, msg.from_user.first_name),
-                     parse_mode="HTML")
+        pass
 
 
 @bot.message_handler(content_types=["left_chat_member"])
@@ -291,9 +294,10 @@ def triggers(msg):
             uids.remove(uid)
     elif msg.chat.id == GROUP_ID_ACTIVE:
         uid_admin = msg.left_chat_member.id
-        cur.execute('DELETE FROM active_admins WHERE uids_admins = %s', [uid_admin])
-        conn.commit()
-        uids_admins.remove(uid_admin)
+        if uid_admin in uids_admins:
+            cur.execute('DELETE FROM active_admins WHERE uids_admins = %s', [uid_admin])
+            conn.commit()
+            uids_admins.remove(uid_admin)
     else:
         bot.send_message(msg.chat.id, text.notmafia.format(msg.from_user.id, msg.from_user.first_name),
                          parse_mode="HTML")
